circular import problem while split visits
